+++ packages/core/directive/attribute.ts
@@ -15,11 +15,28 @@ class AttributeDecorator extends InjectDecoratorFactory {
     super(parameterName, propertyName);
   }
 
+  /**
+   * @Attribute for parameter of method
+   *
+   * @param {ParameterInjectInfoType} injectInfo
+   * @param {Injector} injector
+   * @returns
+   * @memberof AttributeDecorator
+   */
   public resolveParameterInject(injectInfo: ParameterInjectInfoType, injector: Injector) {
     const elementRef: ElementRef = injector.getInstance(ElementRef);
     const renderer: Renderer = injector.getInstance(Renderer);
     return renderer.getAttribute(elementRef.nativeElement, injectInfo.decoratorArgument);
   }
+
+  /**
+   * @Attribute for property of instance
+   *
+   * @param {PropertyInjectInfoType} injectInfo
+   * @param {Injector} injector
+   * @returns
+   * @memberof AttributeDecorator
+   */
   public resolvePropertyInject(injectInfo: PropertyInjectInfoType, injector: Injector) {
     const elementRef: ElementRef = injector.getInstance(ElementRef);
     const renderer: Renderer = injector.getInstance(Renderer);