+++ packages/platform-server/renderer/renderer.ts
@@ -99,7 +99,9 @@ export class PlatfromServerRenderer extends Renderer {
   }
 
   public removeChild(parent: Element, child: Element): void {
-    (parent as Element).removeChild(child as Element);
+    if (parent.contains(child)) {
+      (parent as Element).removeChild(child as Element);
+    }
   }
 
   public appendChild(parent: Element, child: Element): void {